fix: Ensure updates to error list are not lost
Previously, the error list was passed to functions as a slice. The
functions used 'append' to add an error to the slice, but 'append'
creates a new slice, and the caller did not see this update.

The caller did not see the update because the functions did neither of
these:
(a) Return the new slice to the caller
(b) Change the caller's reference to the slice to point to the new
slice.

See https://go.dev/play/p/dTWLE_46Hv7 for a concise example of the bug.

Co-authored-by: Dimitri Koshkin <dimitri.koshkin@gmail.com>